api-python/requirements.txt
# FastAPI
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Database
psycopg2-binary==2.9.10
clickhouse-driver==0.2.9
redis==5.2.0

# Data processing
pandas==2.2.3
numpy==2.1.3

# LangGraph (AI Agent)
langgraph==0.2.45
langchain==0.3.9
langchain-google-genai==2.0.5

# Utils
python-dotenv==1.0.1
python-dateutil==2.9.0
httpx==0.27.0

# Monitoring
prometheus-client==0.21.0
